removed reflection classes from GenericsLib
The reflection classes created were unsafe in that they allow the code to run on different versions of Minecraft without the plugin author verifying the code is still valid.

Replaced with NMS utils and managers. Allows creating NMS code designed for specific NMS versions and registering it with the manager. The manager distributes the code specific to the current NMS version.

If the current NMS version is not valid, the NmsManager disables the plugin.