fix color from usage text bleeds into template color
the color red in error messages is appended as a prefix when the message is displayed. The text formatter does not know about the red when it parses the text so it doesnt know to re-insert red after inserting the usage. The color must be explicitly set in the template to overcome this.