update modules lib to support direct class loading
add ClassLoadMethod enum to specify how the loader is to be used

add IClassNameFactory to JarModuleLoaderSettings so implementing classes can specify which class is the module in a jar file.

moved getDefault* methods from JarModuleLoader to JarModuleLoaderSettings

updated IModuleFactory and IModuleInfoFactory to pass in the JarModuleLoader that is calling the create method. Fixed ProviderLoaderSettings and ScriptEngineLoaderSettings to match.