add Economy provider wrapper to ensure events are called
add Economy provider wrapper to ensure economy related events are called so long as GenericsLib is used to retrieve the provider.

Generics economy events will not fire if the economy provider does not wrap its own classes and a plugin that doesn't use GenericsLib performs an action that would have fired an event.