overhaul data node settings manager
restricted the number of valid value types that can be used

for other value types, a converter can be set on the property definition to convert back and forth between valid types and other types
converters can also be used to convert between human readable words and complex values. i.e. The user types the name of an object, which is converted and stored as the objects unique id.

add SettingsBuilder to make it easier to create PropertyDefinition maps and SettingsManager instance.

update CommandUtils to use