fix floating items duplicate after server restart
suspect unloaded chunks entity data is flushed/cleared before it can be used to remove the entity.

solution:
if a floating item is despawned while GenericsLib is disabled, it is removed like normal except the entity unique id is not wiped from the data node and a dispose flag is set. The entity is removed the next time the manager loads the data node.