add SyncLocation, partially thread safe Location
extends Bukkit Location

Useful when a location needs to be instantiated on an asynchronous thread where it's not safe to retrieve a Bukkit World instance. Retains the name of the world so the world can be retrieved later or the name can be used exclusively.

modified to use SyncLocation: data storage, regions, LocationUtils, NucleusByteReader/Writer

undo Iteration3DTask and IterationTask #onInterationBegin method performed on primary thread always