remove long lived objects on script reload
Objects created by scripts and for scripts are long lived and generally do not get garbage collected after they are released during a reload until the heap is full. This can lead to the assumption that there is a memory leak as memory fills up due to multiple script reloads when in fact there is none. It can also cause a full garbage collection at an inconvenient time.